Revert "More cli."
This reverts commit 29595274dedb53ff37661683604f7cff044fa202.
