+++ packages/core/src/Tool.ts
@@ -21,7 +21,7 @@ import Emitter from './Emitter';
 import ModuleLoader, { Constructor } from './ModuleLoader';
 import Plugin from './Plugin';
 import Reporter from './Reporter';
-import DefaultReporter from './reporters/BoostReporter';
+import DefaultReporter from './reporters/DefaultReporter';
 import ErrorReporter from './reporters/ErrorReporter';
 import enableDebug from './helpers/enableDebug';
 import handleMerge from './helpers/handleMerge';
