+++ packages/core/src/reporters/DefaultReporter.ts
@@ -22,12 +22,12 @@ export default class DefaultReporter extends Reporter<Line> {
     this.console
       .on('start', this.handleStart)
       .on('task', this.handleTask)
-      // .on('task.pass', this.handleTaskComplete)
-      // .on('task.fail', this.handleTaskComplete)
+      .on('task.pass', this.handleTaskComplete)
+      .on('task.fail', this.handleTaskComplete)
       .on('routine', this.handleRoutine)
-      // .on('routine.pass', this.handleRoutineComplete)
-      // .on('routine.fail', this.handleRoutineComplete)
-      // .on('command.data', this.handleCommand)
+      .on('routine.pass', this.handleRoutineComplete)
+      .on('routine.fail', this.handleRoutineComplete)
+      .on('command.data', this.handleCommand)
       .on('render', this.handleRender);
   }
 
@@ -44,6 +44,13 @@ export default class DefaultReporter extends Reporter<Line> {
         ),
       0,
     );
+  }
+
+  /**
+   * Calculate the current number of tasks that have completed.
+   */
+  calculateTaskCompletion(tasks: Task<any>[]): number {
+    return tasks.reduce((sum, task) => (task.hasPassed() || task.isSkipped() ? sum + 1 : sum), 0);
   }
 
   /**
@@ -85,66 +92,67 @@ export default class DefaultReporter extends Reporter<Line> {
     this.keyLength = this.calculateKeyLength(routines);
   };
 
-  // handleCommand = () => {
-  //   this.console.render();
-  // };
+  handleCommand = () => {
+    this.console.render();
+  };
 
   handleTask = (task: Task<any>) => {
-  //   const line = this.findLine(row => row.routine === task.parent);
+    const line = this.findLine(row => row.routine === task.parent);
 
-  //   if (line) {
-  //     line.tasks.push(task);
-  //     this.console.render(this.renderLine(line.routine, task));
-  //   }
-  // };
+    if (line) {
+      line.tasks.push(task);
+    }
 
-  // handleTaskComplete = (task: Task<any>) => {
-  //   const line = this.findLine(row => row.routine === task.parent);
+    this.console.render();
+  };
 
-  //   if (line) {
-  //     line.tasks = line.tasks.filter(t => t !== task);
-  //   }
+  handleTaskComplete = (task: Task<any>) => {
+    const line = this.findLine(row => row.routine === task.parent);
 
-  //   this.console.render();
-  // };
+    if (line) {
+      line.tasks = line.tasks.filter(t => t !== task);
+    }
+
+    this.console.render();
+  };
 
   handleRender = () => {
-    // this.lines.forEach(({ routine, tasks, depth }) => {
-    //   this.renderLine(routine, null, depth);
-    //   tasks.forEach(task => {
-    //     this.renderLine(routine, task, depth);
-    //   });
-    // });
+    this.lines.forEach(({ routine, tasks, depth }) => {
+      this.renderLine(routine, null, depth);
+
+      tasks.forEach(task => {
+        this.renderLine(routine, task, depth);
+      });
+    });
   };
 
   handleRoutine = (routine: Routine<any, any>, value: any, wasParallel: boolean) => {
-    // this.addLine({
-    //   depth: this.depth,
-    //   routine,
-    //   tasks: [],
-    // });
+    this.addLine({
+      depth: this.depth,
+      routine,
+      tasks: [],
+    });
 
-    this.console.render(this.renderLine(routine));
+    this.console.render();
 
     if (!wasParallel) {
       this.depth += 1;
     }
   };
 
-  // handleRoutineComplete = (routine: Routine<any, any>, result: any, wasParallel: boolean) => {
-  //   if (!wasParallel) {
-  //     this.depth -= 1;
-  //   }
+  handleRoutineComplete = (routine: Routine<any, any>, result: any, wasParallel: boolean) => {
+    if (!wasParallel) {
+      this.depth -= 1;
+    }
 
-  //   if (this.depth > 0 && this.tool.config.output < 3) {
-  //     this.removeLine(line => line.routine === routine);
-  //   }
+    if (this.depth > 0 && this.tool.config.output < 3) {
+      this.removeLine(line => line.routine === routine);
+    }
 
-  //   this.console.render();
-  // };
+    this.console.render();
+  };
 
-  renderLine(routine: Routine<any, any>, task: Task<any> | null = null): string {
-    const depth = this.calculateParentDepth(routine);
+  renderLine(routine: Routine<any, any>, task: Task<any> | null, depth: number) {
     const indent = depth * 2;
     const key =
       this.indent(depth) + (task ? '' : routine.key.toUpperCase()).padEnd(this.keyLength - depth);
@@ -174,6 +182,6 @@ export default class DefaultReporter extends Reporter<Line> {
         this.console.size().columns - output.length - status.length,
       ) + status;
 
-    return `${output}\n`;
+    this.console.write(output, 1);
   }
 }
